Converted profile page to be the settings page
